+++ lasher.c
@@ -5,8 +5,8 @@
 #define BYTE unsigned char
 #define WORD unsigned int
 #define SIZE unsigned long long
-#define SW sizeof(WORD)
-#define SS sizeof(SIZE)
+#define SW (int)sizeof(WORD)
+#define SS (int)sizeof(SIZE)
 #define NI 16
 #define NO 4
 #define NS 16
@@ -268,7 +268,7 @@ Rotation is separate from addition to prevent recomputation.
 	size += NI*SW;
 }
 
-static void last(const BYTE ba[NI*SW],size_t nb)
+static void last(const BYTE ba[NI*SW],int nb)
 {
 	int i,j,k,l;
 	BYTE temp[2*NI*SW];
